Rentals and Customer repositories and controllers + rentalModel creation logic( columns relations)

+++ src/main/java/com/api/boardcamp/dtos/rentalDto.java
@@ -11,12 +11,12 @@ public class rentalDto {
     @NotBlank
     @NotNull
     @Positive
-    private Integer customerId;
+    private Long customerId;
 
     @NotBlank
     @NotNull
     @Positive
-    private Integer gameId;
+    private Long gameId;
 
     @NotBlank
     @NotNull
